Fixing syslink and tiler to 27.5 points.
Fixing the syslink and tiler trees to the approved engineering
drop points to enable development on top of these points.  The
generic trees are used for both 24.x and 27.x work so working
from the generic tree could break the 27.x builds frequently.

The manifest will be updated with the new fixed points based
on the approval of Muthu or Saket once testing has been done
and new DOMX changes are ready.